+++ service/collector.go
@@ -183,7 +183,15 @@ LOOP:
 			col.logger.Info("Received shutdown request")
 			break LOOP
 		case <-ctx.Done():
-			col.logger.Error("Context done, terminating process", zap.Error(ctx.Err()))
+			col.logger.Info("Context done, terminating process")
+
+			// Log error for when a context has timed out.
+			// A context that is canceled could be do to intercepted signals or intentional
+			// and we don't want to log an error with a stack trace to the user.
+			err := ctx.Err()
+			if errors.Is(err, context.DeadlineExceeded) {
+				col.logger.Error("Context deadline exceeded", zap.Error(err))
+			}
 			// Call shutdown with background context as the passed in context has been canceled
 			return col.shutdown(context.Background())
 		}